refactor(auth): split actions into separate files
- Split auth-related actions into individual files for better organization
- Move handleLogin to handleLogin.js
- Extract handleLogout to logout.js
- Extract verifySession to verifySession.js
- Update import paths in LoginForm and Navbar components
- Redirect to /events after successful login
