+++ app/actions/verifySession.js
@@ -1,0 +1,5 @@
+"use server";
+
+export const verifySession = async (session) => {
+  return !!session;
+};
